requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
kafka-python==2.0.2
psutil==5.9.6
requests==2.31.0